requirements.txt
langgraph==0.2.34
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.8
chromadb==0.5.5
sentence-transformers==3.0.1
pydantic==2.8.2
python-dotenv==1.0.1
uvloop==0.19.0; platform_system != "Windows"
